Delete old files and rename new ones
This commit finalizes the transition between the previous
analysis routine using only functions and the new one using an
object to run the routine. All of the intermediate files have
been deleted and final routine was renamed to the existing
analysis_reduction.py
